assignment 28 complete

+++ app/controllers/topics_controller.rb
@@ -16,6 +16,10 @@ class TopicsController < ApplicationController
    end
    
    def create
+       if current_user.moderator?
+         flash[:alert] = "You must be an admin to do that."
+       redirect_to topics_path
+     end
      
      @topic = Topic.new(topic_params)
  
@@ -46,6 +50,11 @@ class TopicsController < ApplicationController
    end
    
    def destroy
+     if current_user.moderator?
+         flash[:alert] = "You must be an admin to do that."
+       redirect_to topics_path
+     end
+     
      @topic = Topic.find(params[:id])
  
      if @topic.destroy
@@ -64,7 +73,7 @@ private
    end
    
    def authorize_user
-     unless current_user.admin?
+     unless current_user.admin? || current_user.moderator?
        flash[:alert] = "You must be an admin to do that."
        redirect_to topics_path
      end
